refactor(#3038): extract ConfigSnapshotCache to voice_config_cache sibling module
god-object 분해 다음 묶음. VoiceBargeInRuntime 의 F6 (#2046) Config
snapshot 핫캐시 관심사(config_cache 필드 + 전용 상수 VOICE_CONFIG_CACHE_TTL)를
in-file 정의에서 sibling 모듈 src/services/discord/voice_config_cache.rs 로
물리적으로 격리해 giant production LoC 를 실제로 감소시킨다.

- ConfigSnapshotCache(struct + impl)와 VOICE_CONFIG_CACHE_TTL 을 sibling
  모듈로 이동. 응집도 근거: 단일 Mutex<Option<(Instant, Arc<Config>)>>
  슬롯과 그 TTL 상수만으로 구성된 self-contained 캐시이며 call-site 가
  적다(production 2: lookup_within_ttl/store, test seed 4).
- 동시성 의미 보존: Mutex lock discipline, TTL 폴백(poisoned/contended →
  miss), store/seed 의미가 byte-for-byte 불변. spawn_blocking reload 는
  runtime 에 그대로 남겨 async 제어 흐름/side-effect 순서 불변.
- 가시성: 이동 후 voice_barge_in 및 그 #[cfg(test)] mod tests(super::*)에서
  접근하도록 메서드를 pub(in crate::services::discord) 로 노출. seed 는
  #[cfg(test)] 유지.

LoC: voice_barge_in.rs production 4834 → 4780 (-54, sibling 격리로 실제 감소).
mod.rs 는 모듈 등록 1줄 추가(5185 → 5186). 두 baseline 동기화.

게이트: cargo fmt --check clean / cargo check --lib(신규 경고 0) /
--lib --tests clean / cargo test --lib voice_barge_in 52 passed /
giant_file_ratchet count 0 / freshness passed / ci-script-checks exit 0.

Refs #3038
