+++ ARCHITECTURE.md
@@ -539,6 +539,7 @@ src/
 │   │   ├── voice_acknowledgement.rs
 │   │   ├── voice_background_driver.rs
 │   │   ├── voice_barge_in.rs
+│   │   ├── voice_config_cache.rs
 │   │   ├── voice_routing.rs
 │   │   └── watcher_panel_parity.rs
 │   ├── dispatches/
